+++ x/cronos/client/cli/tx.go
@@ -1,6 +1,7 @@
 package cli
 
 import (
+	"encoding/json"
 	"fmt"
 	"strconv"
 	"strings"
@@ -16,8 +17,13 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/cosmos/cosmos-sdk/client"
-
+	genutilcli "github.com/cosmos/cosmos-sdk/x/genutil/client/cli"
+	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
+	ibcfeetypes "github.com/cosmos/ibc-go/v7/modules/apps/29-fee/types"
 	"github.com/crypto-org-chain/cronos/v2/x/cronos/types"
+	evmtypes "github.com/evmos/ethermint/x/evm/types"
+	feemarkettypes "github.com/evmos/ethermint/x/feemarket/types"
+	gravitytypes "github.com/peggyjv/gravity-bridge/module/v2/x/gravity/types"
 )
 
 // GetTxCmd returns the transaction commands for this module
@@ -37,7 +43,7 @@ func GetTxCmd() *cobra.Command {
 	cmd.AddCommand(CmdUpdateTokenMapping())
 	cmd.AddCommand(CmdTurnBridge())
 	cmd.AddCommand(CmdUpdatePermissions())
-
+	cmd.AddCommand(MigrateGenesisCmd())
 	return cmd
 }
 
@@ -307,5 +313,89 @@ func CmdUpdatePermissions() *cobra.Command {
 	}
 
 	flags.AddTxFlagsToCmd(cmd)
+	return cmd
+}
+
+type ExportEvmGenesisState struct {
+	evmtypes.GenesisState
+	Params ExportEvmParams `json:"params"`
+}
+
+type ExportEvmParams struct {
+	evmtypes.Params
+	ExtraEIPs []string `json:"extra_eips"`
+}
+
+type ExportFeemarketGenesisState struct {
+	feemarkettypes.GenesisState
+	Params   ExportFeemarketParams `json:"params"`
+	BlockGas uint64                `json:"block_gas,string"`
+}
+
+type ExportFeemarketParams struct {
+	feemarkettypes.Params
+	EnableHeight int64 `json:"enable_height,string"`
+}
+
+func Migrate(appState genutiltypes.AppMap, clientCtx client.Context) genutiltypes.AppMap {
+	// Add feeibc with default genesis.
+	if appState[ibcfeetypes.ModuleName] == nil {
+		appState[ibcfeetypes.ModuleName] = clientCtx.Codec.MustMarshalJSON(ibcfeetypes.DefaultGenesisState())
+	}
+	// Add gravity with default genesis.
+	if appState[gravitytypes.ModuleName] == nil {
+		appState[gravitytypes.ModuleName] = clientCtx.Codec.MustMarshalJSON(gravitytypes.DefaultGenesisState())
+	}
+	var evmState ExportEvmGenesisState
+	err := json.Unmarshal(appState[evmtypes.ModuleName], &evmState)
+	if err != nil {
+		panic(err)
+	}
+	data, err := json.Marshal(evmState)
+	if err != nil {
+		panic(err)
+	}
+	appState[evmtypes.ModuleName] = data
+
+	var feemarketState ExportFeemarketGenesisState
+	err = json.Unmarshal(appState[feemarkettypes.ModuleName], &feemarketState)
+	if err != nil {
+		panic(err)
+	}
+	feemarketState.GenesisState.BlockGas = feemarketState.BlockGas
+	data, err = json.Marshal(feemarketState)
+	if err != nil {
+		panic(err)
+	}
+	appState[feemarkettypes.ModuleName] = data
+	return appState
+}
+
+const flagGenesisTime = "genesis-time"
+
+// migrationMap is a map of SDK versions to their respective genesis migration functions.
+var migrationMap = genutiltypes.MigrationMap{
+	"v1.0": Migrate,
+}
+
+// MigrateGenesisCmd returns a command to execute genesis state migration.
+func MigrateGenesisCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "migrate [target-version] [genesis-file]",
+		Short: "Migrate genesis to a specified target version",
+		Long: fmt.Sprintf(`Migrate the source genesis into the target version and print to STDOUT.
+
+Example:
+$ %s migrate v1.0 /path/to/genesis.json --chain-id=cronos_777-1 --genesis-time=2019-04-22T17:00:00Z
+`, version.AppName),
+		Args: cobra.ExactArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return genutilcli.MigrateHandler(cmd, args, migrationMap)
+		},
+	}
+
+	cmd.Flags().String(flagGenesisTime, "", "override genesis_time with this flag")
+	cmd.Flags().String(flags.FlagChainID, "", "override chain_id with this flag")
+
 	return cmd
 }